Named CSV importer converted into an abstract class CSVImporter which does the actual import of CSV data using header string arrays supplied by the derived class. Named and Altium CSV importers changed to reference the new CSVImporter class.
